fix: properly remove old buffer on rename
Currently, renaming a note does not remove the old buffer. This is because:
- After `saveas`, the old buffer stays, but is unlodaded, so `get_named_buffers` will not include it.
- `cur_note_path` is not a string, buf a `Path`, so the check will never succeed

On another hand, `saveas` overrides the alternate file with the old buffer.
This is more a matter of taste, but I find keeping the altfile what it previously was more intuitive, so the user can't accidentally go back to the now deleted buffer.