feat(config): Add different cadence options to daily notes
This adds a new config field for daily notes: `schedule`.

This is because we currently assume that a daily note is only for a
workweek. On a Sunday, if you used `:ObsidianYesterday`, it returns
Friday's note. In Obsidian, this is not the default behavior.

To not break current behavior, we set `workweek` as the default setting
but allow it to be `calendar` (yesterday is yesterday, tomorrow is
tomorrow, etc).